AppOption source file.
Moved the AppOption class to its own source file.
