+++ OfficeToPDF/WordConverter.cs
@@ -597,76 +597,6 @@ namespace OfficeToPDF
             }
             return false;
         }
-        // We want to be able to reset the options in Word so it doesn't affect subsequent
-        // usage
-        private class AppOption
-        {
-            public string Name { get; set; }
-            public Boolean Value { get; set; }
-            public Boolean OriginalValue { get; set; }
-            public int IntValue { get; set; }
-            public int OriginalIntValue { get; set; }
-            protected Type VarType { get; set; }
-            public AppOption(string name, Boolean value, ref Options wdOptions)
-            {
-                try
-                {
-                    Name = name;
-                    Value = value;
-                    VarType = typeof(Boolean);
-                    OriginalValue = (Boolean)wdOptions.GetType().InvokeMember(name, System.Reflection.BindingFlags.GetProperty, null, wdOptions, null);
-
-                    if (OriginalValue != value)
-                    {
-                        wdOptions.GetType().InvokeMember(name, System.Reflection.BindingFlags.SetProperty, null, wdOptions, new Object[] { value });
-                    }
-                }
-                catch
-                {
-                    // We may be setting word options that are not available in the version of word
-                    // being used, so just skip these errors
-                }
-            }
-            public AppOption(string name, int value, ref Options wdOptions)
-            {
-                try
-                {
-                    Name = name;
-                    IntValue = value;
-                    VarType = typeof(int);
-                    OriginalIntValue = (int)wdOptions.GetType().InvokeMember(name, System.Reflection.BindingFlags.GetProperty, null, wdOptions, null);
-
-                    if (OriginalIntValue != value)
-                    {
-                        wdOptions.GetType().InvokeMember(name, System.Reflection.BindingFlags.SetProperty, null, wdOptions, new Object[] { value });
-                    }
-                }
-                catch
-                {
-                    // We may be setting word options that are not available in the version of word
-                    // being used, so just skip these errors
-                }
-            }
-
-            // Allow the value on the options to be reset
-            public void ResetValue(ref Options wdOptions)
-            {
-                if (VarType == typeof(Boolean))
-                {
-                    if (Value != this.OriginalValue)
-                    {
-                        wdOptions.GetType().InvokeMember(Name, System.Reflection.BindingFlags.SetProperty, null, wdOptions, new Object[] { OriginalValue });
-                    }
-                }
-                else
-                {
-                    if (IntValue != OriginalIntValue)
-                    {
-                        wdOptions.GetType().InvokeMember(Name, System.Reflection.BindingFlags.SetProperty, null, wdOptions, new Object[] { OriginalIntValue });
-                    }
-                }
-            }
-        }
 
         // Remove fill-in fields which can cause blocking dialogs
         private static void RemoveFillInFields(Microsoft.Office.Interop.Word.Document doc, bool updateField)
